refactor(sharing-dialog): move access list to separate dir
