+++ components/sharing-dialog/src/access-list/__tests__/sharing-list-item-icon.test.js
@@ -7,7 +7,7 @@ import {
     SHARE_TARGET_PUBLIC,
     SHARE_TARGET_USER,
     SHARE_TARGET_GROUP,
-} from '../sharing-constants.js'
+} from '../../sharing-constants.js'
 import { SharingListItemIcon } from '../sharing-list-item-icon.js'
 
 describe('<SharingListItemIcon />', () => {
