+++ components/sharing-dialog/src/access-list/index.js
@@ -1,0 +1,1 @@
+export { AccessList } from './access-list.js'
