+++ components/sharing-dialog/src/access-list/sharing-list-item-icon.js
@@ -8,7 +8,8 @@ import {
     SHARE_TARGET_PUBLIC,
     SHARE_TARGET_USER,
     SHARE_TARGET_GROUP,
-} from './sharing-constants.js'
+} from '../sharing-constants.js'
+import { IconInitials } from './icon-initials.js'
 
 export const SharingListItemIcon = ({ target, name }) => {
     switch (target) {
