+++ components/sharing-dialog/src/access-list/sharing-list-item.js
@@ -2,9 +2,9 @@ import { Divider } from '@dhis2-ui/divider'
 import { colors } from '@dhis2/ui-constants'
 import PropTypes from 'prop-types'
 import React from 'react'
-import { AccessSelect } from './access-select.js'
-import { isRemovableTarget } from './helpers.js'
-import i18n from './locales/index.js'
+import { AccessSelect } from '../access-select.js'
+import { isRemovableTarget } from '../helpers.js'
+import i18n from '../locales/index.js'
 import { SharingListItemContext } from './sharing-list-item-context.js'
 import { SharingListItemIcon } from './sharing-list-item-icon.js'
